moved types to sub directory
