--- wls-gui-wahllokalsystem/src/types/ergebnismeldung/MBW/BedenklicherStimmzettel.ts
@@ -1,8 +1,0 @@
-import type { SupplementEnum } from "@/types/ergebnismeldung/MBW/SupplementEnum.ts";
-import type { ValidityEnum } from "@/types/ergebnismeldung/MBW/ValidityEnum.ts";
-
-export interface BedenklicherStimmzettel {
-  orderIndex: number;
-  supplements: SupplementEnum[];
-  validity: ValidityEnum;
-}
