+++ wls-gui-wahllokalsystem/src/types/ergebnismeldung/MBW/bedenklicheStimmzettel/SupplementEnum.ts
