+++ wls-gui-wahllokalsystem/src/types/ergebnismeldung/MBW/bedenklicheStimmzettel/ValidityEnum.ts
