+++ wls-gui-wahllokalsystem/tests/composables/ergebnismeldung/MBW/bedenklicherStimmzettelMapper.spec.ts
@@ -1,5 +1,5 @@
 import type { BedenklicherStimmzettelDTO } from "@/api/wls-clients/generated-ergebnismeldung-api";
-import type { BedenklicherStimmzettel } from "@/types/ergebnismeldung/MBW/BedenklicherStimmzettel.ts";
+import type { BedenklicherStimmzettel } from "@/types/ergebnismeldung/MBW/bedenklicheStimmzettel/BedenklicherStimmzettel.ts";
 
 import { useBedenklicherStimmzettelTestDataFactory } from "@tests/utils/ergebnismeldung/MBW/bedenklicherStimmzettelTestDataFactory.ts";
 import { beforeEach, describe, expect, it } from "vitest";
@@ -9,8 +9,8 @@ import {
   BedenklicherStimmzettelDTOValidityEnum,
 } from "@/api/wls-clients/generated-ergebnismeldung-api";
 import { useBedenklicherStimmzettelMapper } from "@/composables/ergebnismeldung/MBW/bedenklicherStimmzettelMapper.ts";
-import { SupplementEnum } from "@/types/ergebnismeldung/MBW/SupplementEnum.ts";
-import { ValidityEnum } from "@/types/ergebnismeldung/MBW/ValidityEnum.ts";
+import { SupplementEnum } from "@/types/ergebnismeldung/MBW/bedenklicheStimmzettel/SupplementEnum.ts";
+import { ValidityEnum } from "@/types/ergebnismeldung/MBW/bedenklicheStimmzettel/ValidityEnum.ts";
 
 const { prepareBedenklicherStimmzettelDTO, prepareBedenklicherStimmzettel } =
   useBedenklicherStimmzettelTestDataFactory();
