+++ wls-gui-wahllokalsystem/tests/utils/ergebnismeldung/MBW/bedenklicherStimmzettelTestDataFactory.ts
@@ -1,5 +1,5 @@
 import type { BedenklicherStimmzettelDTO } from "@/api/wls-clients/generated-ergebnismeldung-api";
-import type { BedenklicherStimmzettel } from "@/types/ergebnismeldung/MBW/BedenklicherStimmzettel.ts";
+import type { BedenklicherStimmzettel } from "@/types/ergebnismeldung/MBW/bedenklicheStimmzettel/BedenklicherStimmzettel.ts";
 import type { Builder } from "@tests/utils/Builder.ts";
 
 import { proxyBuilder } from "@tests/utils/Builder.ts";
@@ -9,8 +9,8 @@ import {
   BedenklicherStimmzettelDTOSupplementsEnum,
   BedenklicherStimmzettelDTOValidityEnum,
 } from "@/api/wls-clients/generated-ergebnismeldung-api";
-import { SupplementEnum } from "@/types/ergebnismeldung/MBW/SupplementEnum.ts";
-import { ValidityEnum } from "@/types/ergebnismeldung/MBW/ValidityEnum.ts";
+import { SupplementEnum } from "@/types/ergebnismeldung/MBW/bedenklicheStimmzettel/SupplementEnum.ts";
+import { ValidityEnum } from "@/types/ergebnismeldung/MBW/bedenklicheStimmzettel/ValidityEnum.ts";
 
 const { generateRandomNumber, getRandomItem } = useCommonTestDataFactory();
